feat(UHeader): Add download app button
This commit adds a "Download app" button to the UHeader component. The
button is displayed only when the build mode is not set to "electron".
Clicking on the button navigates the user to the latest release page on
GitHub. This addition provides a convenient way for users to download
the app directly. The FontAwesome library was also updated to include
the "faCloudArrowDown" icon needed for the button.