refactor(ArtistDetailView): Add padding to the bottom of the container
This change adds a `pb-32` class to the main container in the
`ArtistDetailView.vue` component. Previously, there was no bottom
padding, causing the page content to be cut off at the bottom. This
change ensures that there is enough space for the content, providing a
better user experience. No issues are associated with this change.